use default implementation for import_table()
