+++ src/any.rs
@@ -278,18 +278,6 @@ impl DbQuery for AnyPool {
         }
     }
 
-    /// Implements [DbQuery::import_table()]
-    async fn import_table(&self, tsv: &str) -> Result<(), DbError> {
-        match self {
-            #[cfg(feature = "rusqlite")]
-            AnyPool::Rusqlite(pool) => pool.import_table(tsv).await,
-            #[cfg(feature = "tokio-postgres")]
-            AnyPool::TokioPostgres(pool) => pool.import_table(tsv).await,
-            #[cfg(feature = "libsql")]
-            AnyPool::LibSQL(pool) => pool.import_table(tsv).await,
-        }
-    }
-
     /// Implements [DbQuery::drop_table()]
     async fn drop_table(&self, table: &str) -> Result<(), DbError> {
         match self {
