+++ src/rusqlite.rs
@@ -7,7 +7,7 @@ use crate::{
     db_kind::{DbKind, MAX_PARAMS_SQLITE, SQLiteKind},
     db_value::{DbParams, DbRow, DbRows, DbValue, IntoDbParams, IntoDbRows, JsonValue},
     parse::validate_table_name,
-    shared::{self, EditType, edit},
+    shared::{EditType, edit},
 };
 use deadpool_sqlite::{
     Config, Pool, Runtime,
@@ -437,11 +437,6 @@ impl DbQuery for RusqlitePool {
             returning,
         )
         .await
-    }
-
-    /// Implements [DbQuery::import_table()] for SQLite.
-    async fn import_table(&self, tsv: &str) -> Result<(), DbError> {
-        shared::import_table(self, tsv).await
     }
 
     /// Implements [DbQuery::drop_table()] for SQLite.
